Initial Commit
Bhagavat Gita (The Song by God) is considered one of the most important and widely discussed religious
scripture of Hinduism. It is 700 verse scripture which contains dialogues between Prince and
disciple Arjuna and his mentor and guide Krishna, known as a text that summarizes and comprises
the core of Hindu philosophy. In the Mahabharata war, Arjuna, the most skilled and prominent
archer and Krishna as his charioteer and other Pandavas were at war with their cousins Kauravas.
Arjuna faced a moral dilemma where he did not want to go to war with his relatives and wanted to
renounce everything to become a Yogi. As his mentor and guide, Lord Krishna reminded Arjuna
of his Kshatriya duty and taught him the concepts of a karma yogi.
Although the Bhagavad Gita might be thousands of years old, the philosophy, logic, and knowledge
are timeless. Understanding the appropriate meaning and interpretation of slokas in one’s
native language can helps in gaining a deeper understanding of the hows and whys of everyday life.
During the tough phases of life, most of us wish to refer to Bhagavad Gita for the solutions.
However, understanding the divine concepts and language is not very easy and requires broad
knowledge. So this motivated us to work on the semantic meaning of the Bhagavad Gita which will
then provide the solution for any problem stated by the user using teachings of the Lord Krishna.